toys/confidence_intervals/requirements.txt
Flask==3.0.0
pywebview==4.4.1
pyinstaller==6.3.0
